+++ cmd/enable.go
@@ -18,9 +18,10 @@ var enableCmd = &cobra.Command{
 }
 
 var enableTunnelCmd = &cobra.Command{
-	Use:   "tunnel [profile]",
-	Short: "Enable a tunnel in a profile",
-	Args:  cobra.ExactArgs(1),
+	Use:               "tunnel [profile]",
+	Short:             "Enable a tunnel in a profile",
+	Args:              cobra.ExactArgs(1),
+	ValidArgsFunction:  profileNameCompletionFunc,
 	Run: func(cmd *cobra.Command, args []string) {
 		profileName := args[0]
 		name, _ := cmd.Flags().GetString("name")